+++ SoftMeter-C-Api.h
@@ -20,11 +20,24 @@
 	typedef		char				smChar_t;
 #endif
 
+/*
+	Note for Window's DLLs, and the calling convention.
+	
+	The following functions use the __cdecl calling convention.
+	Inside the DLL, they have exact counterparts that use the __stdcall calling convention.
+	The function names of their counterparts have a suffix of "_stdcall", e.g. 
+		start() // this is the __cdecl function
+		start_stdcall() // this is the __stdcall function
+	From your Windows application you can call the set you prefer.
+	
+	In other words, the DLL contains both calling conventions.
+*/
+
 // get the version string of the library. 
-EXPORT_API const smChar_t*	getVersion(void);
+EXPORT_API const smChar_t* getVersion(void);
 
 // get the full path name of the log file. 
-EXPORT_API const smChar_t*	getLogFilename(void);
+EXPORT_API const smChar_t* getLogFilename(void);
 
 // Enable or disable the log file. The default is disabled.
 // - Parameter macBundleId is used only in Mac OS. Under Windows you can pass NULL.
@@ -74,4 +87,22 @@ EXPORT_API bool sendEvent(const smChar_t *eventAction, const smChar_t *eventLabe
 EXPORT_API bool sendException(const smChar_t *exceptionDescription, const bool isFatal);
 
 
+#ifdef _WIN32
+	// __stdcall version of all the functions 
+	// The function names are appended with _stdcall
+	/* This section is REMed because these functions are exported by a .DEF file 
+	EXPORT_API const smChar_t*	__stdcall getVersion_stdcall(void);
+	EXPORT_API const smChar_t*	__stdcall getLogFilename_stdcall(void);
+	EXPORT_API void __stdcall enableLogfile_stdcall(const smChar_t *appName, const smChar_t *macBundleId);
+	EXPORT_API void __stdcall disableLogfile_stdcall(void);
+	EXPORT_API bool __stdcall start_stdcall(const smChar_t *appName, const smChar_t *appVersion,
+		const smChar_t *appLicense, const smChar_t *appEdition,
+		const smChar_t *propertyID, const bool  userGaveConsent);
+	EXPORT_API void __stdcall stop_stdcall(void);
+	EXPORT_API bool __stdcall sendPageview_stdcall(const smChar_t *pagePath, const smChar_t *pageTitle);
+	EXPORT_API bool __stdcall sendScreenview_stdcall(const smChar_t *screenName);
+	EXPORT_API bool __stdcall sendEvent_stdcall(const smChar_t *eventAction, const smChar_t *eventLabel, const int eventValue);
+	EXPORT_API bool __stdcall sendException_stdcall(const smChar_t *exceptionDescription, const bool isFatal);
+	*/
 
+#endif